backend/requirements.txt
fastapi==0.115.5
uvicorn==0.34.0
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.0
sqlalchemy==2.0.35
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
requests==2.32.3
pytest==8.3.4
httpx==0.28.1
pytest-asyncio==0.25.0
greenlet==3.1.1